+++ src/components/KInput/KInput.vue
@@ -33,9 +33,10 @@
 
       <input
         v-bind-once="{ id: inputId, ...(helpText && { 'aria-describedby': helpTextId }) }"
-        v-bind="modifiedAttrs"
         :aria-invalid="error || hasError || charLimitExceeded ? 'true' : undefined"
         class="input"
+        :type="type"
+        v-bind="modifiedAttrs"
         :value="getValue()"
         @input="handleInput"
       >
@@ -125,6 +126,11 @@ const props = defineProps({
 
       return true
     },
+  },
+  type: {
+    type: String,
+    required: false,
+    default: 'text',
   },
 })